Estimate a survival function under current status sampling

Estimate a survival function under current status sampling

Usage

currstatCIR(
  time,
  event,
  W,
  SL_control = list(SL.library = c("SL.mean"), V = 5, method = "method.NNLS"),
  HAL_control = list(n_bins = c(5, 10), grid_type = c("equal_range", "equal_mass"), V =
    5),
  deriv_method = "m-spline",
  missing_method = "extended",
  eval_region,
  n_eval_pts = 101
)

Arguments

time

n x 1 numeric vector of observed monitoring times

event

n x 1 numeric vector of status indicators of whether an event was observed prior to the monitoring time.

W

Dataframe of covariates

SL_control

List of Super Learner control parameters

HAL_control

List of haldensify control parameters

deriv_method

Method for computing derivative

missing_method

Method for handling nonresponse (extended CIR vs. complete case, just for testing)

eval_region

Region over which to estimate the survival function

n_eval_pts

Number of points in grid on which to evaluate survival function

Value

data frame giving results