survML: Tools for Flexible Survival Analysis Using Machine Learning

Statistical tools for analyzing time-to-event data using machine learning. Implements survival stacking for conditional survival estimation, isotonic regression for current status data, and methods for algorithm-agnostic variable importance. See Wolock CJ, Gilbert PB, Simon N, and Carone M (2024) doi:10.1080/10618600.2024.2304070 .

See also

Author

Maintainer: Charles Wolock cwolock@gmail.com (ORCID) [copyright holder]

Other contributors:

Avi Kenny avi.kenny@gmail.com (ORCID) [contributor]